// lib/widgets/types.ts
/*
 * @Author: your name
 * @Date: 2021-01-09 16:09:14
 * @LastEditTime: 2021-01-11 11:38:04
 * @LastEditors: Please set LastEditors
 * @Description: In User Settings Edit
 * @FilePath: \vue3-component\vue3-json-schema-form\lib\widgets\types.ts
 */

export interface OptionsDefine {
  label: string
  value: any
}